+++ templates/standard/.eleventy/utils/extension_of.js
@@ -2,6 +2,9 @@ const path = require('path');
 
 /**
  * Extracts extension name from url.
+ * By default, it is without leading point character.
  * @param {string} url
+ * @param {boolean} [withDot]
  */
-module.exports.extensionOf = (url) => path.extname(url).slice(1);
+module.exports.extensionOf = (url, withDot = false) =>
+  withDot ? path.extname(url) : path.extname(url).slice(1);